refactor(app): move the majority of the bootstrapping to global
