+++ core/src/components/app/app.tsx
@@ -1,12 +1,9 @@
 import type { ComponentInterface } from '@stencil/core';
 import { Build, Component, Element, Host, Method, h } from '@stencil/core';
 import type { FocusVisibleUtility } from '@utils/focus-visible';
-import { shouldUseCloseWatcher } from '@utils/hardware-back-button';
-import { printIonWarning } from '@utils/logging';
-import { isPlatform } from '@utils/platform';
 
 import { config } from '../../global/config';
-import { getIonTheme } from '../../global/ionic-global';
+import { getIonTheme, rIC } from '../../global/ionic-global';
 
 /**
  * @virtualProp {"ios" | "md"} mode - The mode determines the platform behaviors of the component.
@@ -22,43 +19,9 @@ export class App implements ComponentInterface {
   @Element() el!: HTMLElement;
 
   componentDidLoad() {
+    // TODO can we move focus visible to global
     if (Build.isBrowser) {
       rIC(async () => {
-        const isHybrid = isPlatform(window, 'hybrid');
-        if (!config.getBoolean('_testing')) {
-          import('../../utils/tap-click').then((module) => module.startTapClick(config));
-        }
-        if (config.getBoolean('statusTap', isHybrid)) {
-          import('../../utils/status-tap').then((module) => module.startStatusTap());
-        }
-        if (config.getBoolean('inputShims', needInputShims())) {
-          /**
-           * needInputShims() ensures that only iOS and Android
-           * platforms proceed into this block.
-           */
-          const platform = isPlatform(window, 'ios') ? 'ios' : 'android';
-          import('../../utils/input-shims/input-shims').then((module) => module.startInputShims(config, platform));
-        }
-        const hardwareBackButtonModule = await import('../../utils/hardware-back-button');
-        const supportsHardwareBackButtonEvents = isHybrid || shouldUseCloseWatcher();
-        if (config.getBoolean('hardwareBackButton', supportsHardwareBackButtonEvents)) {
-          hardwareBackButtonModule.startHardwareBackButton();
-        } else {
-          /**
-           * If an app sets hardwareBackButton: false and experimentalCloseWatcher: true
-           * then the close watcher will not be used.
-           */
-          if (shouldUseCloseWatcher()) {
-            printIonWarning(
-              'experimentalCloseWatcher was set to `true`, but hardwareBackButton was set to `false`. Both config options must be `true` for the Close Watcher API to be used.'
-            );
-          }
-
-          hardwareBackButtonModule.blockHardwareBackButton();
-        }
-        if (typeof (window as any) !== 'undefined') {
-          import('../../utils/keyboard/keyboard').then((module) => module.startKeyboardAssist(window));
-        }
         import('../../utils/focus-visible').then((module) => (this.focusVisible = module.startFocusVisible()));
       });
     }
@@ -94,33 +57,3 @@ export class App implements ComponentInterface {
     );
   }
 }
-
-const needInputShims = () => {
-  /**
-   * iOS always needs input shims
-   */
-  const needsShimsIOS = isPlatform(window, 'ios') && isPlatform(window, 'mobile');
-  if (needsShimsIOS) {
-    return true;
-  }
-
-  /**
-   * Android only needs input shims when running
-   * in the browser and only if the browser is using the
-   * new Chrome 108+ resize behavior: https://developer.chrome.com/blog/viewport-resize-behavior/
-   */
-  const isAndroidMobileWeb = isPlatform(window, 'android') && isPlatform(window, 'mobileweb');
-  if (isAndroidMobileWeb) {
-    return true;
-  }
-
-  return false;
-};
-
-const rIC = (callback: () => void) => {
-  if ('requestIdleCallback' in window) {
-    (window as any).requestIdleCallback(callback);
-  } else {
-    setTimeout(callback, 32);
-  }
-};
